+++ app/src/main/java/to/bitkit/viewmodels/AppViewModel.kt
@@ -382,6 +382,8 @@ class AppViewModel @Inject constructor(
                     is Event.PaymentForwarded -> Unit
                     is Event.PaymentReceived -> handlePaymentReceived(event)
                     is Event.PaymentSuccessful -> handlePaymentSuccessful(event)
+                    is Event.ProbeFailed -> Unit
+                    is Event.ProbeSuccessful -> Unit
                     is Event.SpliceFailed -> Unit
                     is Event.SplicePending -> Unit
                     is Event.SyncCompleted -> handleSyncCompleted()